projects/03_coffee_shop_full_stack/starter_code/backend/requirements.txt
astroid==2.2.5
Click==7.0
cryptography==3.2.1
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pycryptodome==3.3.1
pyjwt==1.7.1
pylint==2.3.1
python-jose-cryptodome==1.3.2
python-jose==3.2.0
six==1.12.0
SQLAlchemy==1.3.3
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1